feat: Add BaseTemplate component
The `BaseTemplate` component is added to provide a base layout for the application. It includes a `Navbar` component at the top, followed by the main content wrapped in a `div`, and a `Footer` component at the bottom. This component will be used as a template for other pages in the application.